Added Timers Context Provider
